requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==4.2.7
django-cors==0.1
django-cors-headers==4.2.0
django-filter==23.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
executing==1.2.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.6
whitenoise==6.6.0
